Create a DataDog Event for each Destroy
This change sends an Event to DataDog each time a destroy is called in
Chaos Lemur. It adds a name property to the group property already
present in Member to make the Event more informative. It adds tests for
the new functionality and adds the name property to the existing tests
as needed.

It also reduces the memory requirement for Chaos Lemur.

[#76769358]