Add AWS infrastructure
Previously the only infrastructure was a ‘dummy’ one as a proof of
concept. This change adds a working infrastructure for AWS. This will
be used if the environment variable AWS_ACCESSKEYID is set.

It modifies the definition of a member to include an ID that can be
passed back to the infrastructure to limit the number of lookup calls
required.

It adds tests for the new infrastructure, and updates existing tests for
the new member definition.

It adds a date/time stamp to the stdout logging configuration.

It also updates to the latest version of Spring Boot, and tidies up the
StandardDataDogTest class.

[#76769374]