Add a NoOp DataDog
Previously all Events were logged to DataDog, including for runs on development
machines. This causes noise in the DataDog record. This change adds a NoOp
DataDog that is used if the DATADOG_APIKEY environment variable is not set.

It also tightens the visibility of several elements of code, and updates
documentation accordingly.

[#76867400]