Add state persistence
Previously Chaos Lemur didn't persist its state across restarts. If, for
example, Chaos Lemur was paused and the PaaS forced a reboot, Chaos
Lemur would restart and begin destroying VMs because it had no knowledge
of its former state.

This change adds persistence using a Redis instance to store state. It
reorganizes the code extensively to allow for a RedisStateProvider
along with a 'dummy' SimpleStateProvider. The RedisStateProvider is used
if a necessary environment variable is set, which would typically be the
case when it's run in a PaaS, otherwise the SimpleStateProvider is used
(which allows for local operation).

It temporarily removes the ability to start an ad hoc destroy.

It also updates several dependencies.

[#79007116]