Add a destroy endpoint task
Previously requests for an ad hoc destroy were received without
providing feedback to the caller beyond an 'accepted' response. This
change adds a task url to the response, and a means of querying that
task url to show the state of the task. It also provides an endpoint
to list all tasks.

In addition it changes the download location it uses for the java
buildpack.

[#78625656]