Limit chaos to a deployment
Previously, Chaos Lemur attempted to hide the fact that it would only
run against a BOSH-deployed installation.  This meant that the ability
to blacklist/change probability for a deployment (in addition to job)
was difficult to implement.  This change abandons the pretense that
Chaos Lemur will work against anything other than a BOSH-deployed
installation by changing references to Groups to Jobs and exposing the
Deployment that a VM belongs to.  In addition, the fate engine will
now evaluate the deployment when looking at blacklists and configuring
probability.  Finally, this change added the ability to configure
Chaos Lemur to work as expected except that no VMs would actually be
killed.