vSphere Infrastructure
This change adds support for vSphere to the Infrastructure
abstraction. I found that BOSH doesn't tag vSphere VMs the way it does
for AWS so getting the list of  candidate members required going to
the Director.  Rather than handle this detection differently for the
AWS and vSphere Infrastructures, I refactored the AWS implementation
to do detection the same way.

[#86906990]