Arbitrary DataDog tags
In preparation for running the stream tests simultaneously on two
infrastructures, the ability to contribute arbitrary tags to the
DataDog event has been added.  This will allow us to tag the DataDog
events with something like "infrastructure:aws" or
"infrastructure:vsphere" and then filter in the DataDog dashboards by
those values.

[#87605100]