Remove vSphere VM Destruction
Previously, the vSphere infrastructure both stopped and destroyed the
VM it was working with.  A side-effect of the destruction behavior was
that persistent disks that were attached were also destroyed.  Since
BOSH cannot recover from this situation, this wasn't the correct
behavior for Chaos Lemur.

A possible solution to this was to detach the persistent disks from
the VM first, and then destroy it.  According to the API
documentation, this would cause the VM's files to be removed from the
datastore while leaving the persistent disk files in place.  However,
according to the BOSH team, when a VM is resurrected it will delete
any existing VM files (preserving the persistent disks) automatically,
preventing a leak in the datastore.

Since determining what disks to detach is somewhat time consuming,
this change simply removes the VM destruction step, depending on BOSH
to do the appropriate cleanup.  If it turns out that this is
insufficient (i.e. BOSH doesn't clean up properly), we'll come back
and make vSphere VM destruction a bit more sophisticated.