Improve Infrastructure Choice
Previously if neither AWS nor VSPHERE infrastructures were chosen Chaos
Lemur would default to a built-in ‘simple infrastructure’. This was
potentially confusing for users, as it would result in a running system
that logged the destruction of apparently bogus VMs (actually the
test VMs in the simple infrastructure).

This change requires a deliberate choice of either AWS, VSPHERE, or
SIMPLE_INFRASTRUCTURE to remove this confusion.