Protection against invalid data
Previously either a Member was being created with a null id or when
the non-null id of a Member was searched for, a null VirtualMachine
was being returned.  The fact that we could not determine exactly what
was happening indicated that the servicability around this behavior
was not sufficient.

This change updates the code to be a bit more protective against null
values by ensuring (using Spring assertions) that nulls aren't being
passed where they shouldn't be.  It is quite likely that this will not
actually solve the problems we are seeing, but it will improve the
servicability of the code allowing us to find the problems.