One-time Use vSphere Connections
Previously, Chaos Lemur maintained a single long-lived connection to
the vSphere infrastructure if it was configured to use vSphere.  The
connection itself is HTTP-based, so the socket is not kept open, but
since the payload uses the SOAP protocol, there is a session token. Up
until recently, every use of Chaos Lemur ensured calls to the vSphere
infrastructure happened before the token expired, thus refreshing it
and masking the possibility of expiration causing a failure.  This
failure manifested itself as an empty payload being returned and no
candidate members being available for destruction.

This change introduces a level of indirection to ensure that a new
connection is made (and a new session token created) to the vSphere
infrastructure each time a connection needs to be made.  This isn't
the ideal solution, but given that it is impossible to predict the
schedule that a Chaos Lemur instance will execute on and the
configured session token timeout in vSphere, this is the safest
design.