Add a WHITELIST Option
Previously the only way to manage which VMs would be included or
excluded from destruction was to use the BLACKLIST environment variable.
This was adequate for many cases, but in some cases could lead to
complex and difficult to maintain BLACKLISTs.

This change adds a WHITELIST option that adds in entire deployments or
jobs to be considered for destruction. Combining the BLACKLIST and
WHITELIST provides greater control over destruction candidates.

[#96429730][resolves #1]