Event Type
Previously, Events were reported by passing a title and message that
was created by the Deployer.  However, this more or less only worked
for systems that expected and allowed arbitrary text payloads for
reports.  This was not the case for systems where data needed to be
reported in a more structured way.

This change modifies the Reporter interface to take an Event type.
This type encapsulates the identifier and list of members that were
destroyed for a particular event.  This encapsulation allows the
alternate implementations that require more structured data (e.g.
JDBC) to handle event reports properly.  The existing title/message
implementation was migrated from the Deployer to the DataDogReporter.
A similar, but simpler, implementation was added to the NoOpReporter.