Add support for AWS regions
Previously Chaos Lemur implicitly assumed that it was operating in the
us-east-1 region, which was an unnecessary restriction on users. This
change allows the user to specify the required region, with an explicit
default of us-east-1.

It also updates to the latest versions of the AWS and Spring Boot
dependencies.

[resolves #7]